# contactapi/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "STIS Contact API",
    "description": "API ini menyediakan layanan untuk mengelola dan mengakses informasi kontak yang dimiliki oleh mahasiswa dan dosen di Politeknik Statistika STIS. Layanan ini dirancang untuk mempermudah pencarian informasi kontak dengan menyediakan berbagai endpoint untuk menambah, memperbarui, dan menghapus kontak dari mahasiswa dan dosen. API ini juga mendukung pengelompokan kontak berdasarkan organisasi mahasiswa atau mata kuliah yang diampu oleh dosen serta menawarkan fitur pencarian yang fleksibel. \n\nFitur utama meliputi: \n- Manajemen kontak mahasiswa dan dosen \n- Pengelompokan kontak berdasarkan organisasi atau jurusan \n- Dukungan endpoint untuk CRUD (Create, Read, Update, Delete) pada data kontak \n- Fitur pencarian kontak berdasarkan organisasi atau mata kuliah yang diampu oleh dosen\n\n",
    "contact": {
      "name": "Dodi Firmansyah",
      "url": "https://github.com/dodfrm",
      "email": "222212572@stis.ac.id"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/subjects/{id}": {
      "get": {
        "tags": ["Subject"],
        "summary": "Get a subject by ID",
        "operationId": "getSubjectById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Subject not found"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Found the subject",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Subject"],
        "summary": "Update an existing subject",
        "operationId": "updateSubject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "subjectName": "Kalkulus"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Subject updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Subject not found"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Subject"],
        "summary": "Delete a subject by ID",
        "operationId": "deleteSubject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Subject not found"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Subject deleted"
          }
        }
      }
    },
...
    "/api/subjects": {
      "get": {
        "tags": ["Subject"],
        "summary": "Get all subjects",
        "operationId": "getAllSubjects",
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Found the subjects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Subject"],
        "summary": "Create a new subject",
        "operationId": "createSubject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "subjectName": "Kalkulus"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Subject created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subjectName": {
            "type": "string"
          }
        }
      },
      "ContactSubject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          }
        }
      },
      "ContactOrganization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kelas": {
            "type": "string"
          },
          "jabatan": {
            "type": "string",
            "enum": ["KETUA", "WAKIL_KETUA", "BENDAHARA", "PJ_MATKUL"]
          },
          "periodeJabatan": {
            "type": "string"
          }
        }
      },
...
      "ContactDto": {
        "required": ["fullName", "phone"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contactType": {
            "type": "string",
            "enum": ["MAHASISWA", "DOSEN"]
          },
          "contactOrganizations": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactOrganization"
            }
          },
          "contactSubjects": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactSubject"
            }
          }
        }
      },
      "SubjectDto": {
        "required": ["subjectName"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subjectName": {
            "type": "string"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Input token sebagai 'Bearer \u003Ctoken\u003E'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}